magic-functions: update to new model definition
