+++ examples/magic-functions/src/editor/data/env.ts
@@ -1,0 +1,5 @@
+export default `
+// hacks to remove undefined and globalThis from code suggest
+// https://github.com/microsoft/monaco-editor/issues/2018
+declare module undefined 
+`;
